services/listener/requirements.txt
slack_sdk
google-api-python-client
google-auth
requests
emoji
